amdxdna/ve2: add ve2_hw layer and keep cert firmware load in L2
Restructure the auxiliary VE2 path so device firmware and the hardware
abstraction sit in ve2_hw.c/ve2_hw.h (Layer 2), while ve2_aux.c/h stay
focused on amdxdna_dev_info, ve2_ops, and probe wiring (Layer 1).

ve2_hw.h defines UAPI-related firmware types (e.g. ve2_firmware_version,
ve2_firmware_status, cert memory offsets), the public HAL entry points
ve2_hw_load_cert_firmware() and ve2_hw_init_fw_status_slots(), and
struct ve2_hw_ops with ve2_hw_get_ops() for context and command paths.
Layer 1 must use those interfaces instead of calling the AIE/partition
APIs for firmware.

ve2_hw.c implements cert image load (request_firmware, temporary buffer,
aie partition request/initialize, aie_load_cert_broadcast, version read
from program data), per-column firmware status slot allocation, and
MODULE_FIRMWARE for the cert blob. Default ve2_hw_ops keep ctx_init/
ctx_fini and cmd_submit/cmd_wait as minimal stubs until the downstream
port continues.

ve2_aux.h pulls in ve2_hw.h for the dev handle’s fw_version and fw_slots
members; the migration map comment in the header documents the intended
split: ve2_of → ve2_aux, ve2_mgmt+ve2_fw → ve2_hw, hwctx/debug/resolver
in Layer 1 with calls only through ve2_hw_get_ops() and the new device
firmware helpers.

This matches the VE2 upstream proposal: clear L1 (framework) vs L2
(AIE HAL) boundary and a single place to grow partition and handshake
logic.
